ajuste final

+++ app/Http/Controllers/PublicacionesController.php
@@ -128,7 +128,6 @@ class PublicacionesController extends Controller
             }
 
             try {
-
                 $validacionUnidadProducto = Precios::where([
                     ['producto_id', $datos['idproductos']],
                     ['unidades_id', $datos['idunidades']],
@@ -168,9 +167,10 @@ class PublicacionesController extends Controller
                             $nuevaImagenP->ruta = $imagenPredeterminada;
                             $nuevaImagenP->publicaciones_id = $nuevoPublicacion->id;
                             $nuevaImagenP->save();
-                        } else {
-                            $aErrores[] = '- Escoja al menos una imagen para la publicación o seleccione un producto con imagen predeterminada.';
                         }
+                        // else {
+                        //     $aErrores[] = '- Escoja al menos una imagen para la publicación o seleccione un producto con imagen predeterminada.';
+                        // }
                     } else {
                         if (isset($datos['imagen']) && is_array($datos['imagen'])) {
                             foreach ($datos['imagen'] as $imagen) {
@@ -229,88 +229,127 @@ class PublicacionesController extends Controller
             if ($datos['precio'] == "") {
                 $aErrores[] = '- Añada el precio de venta';
             }
-            if (empty($_FILES['imagen']['name'][0])) {
-                $aErrores[] = '- Escoja al menos una imagen para la publicación';
-            }
+            // if (empty($_FILES['imagen']['name'][0])) {
+            //     $aErrores[] = '- Escoja al menos una imagen para la publicación';
+            // }
             if (count($aErrores) > 0) {
                 throw new \Exception(join('</br>', $aErrores));
             }
             try {
 
-                $validacion = Publicaciones::where([
+
+                $validacionPublicacionExiste = Publicaciones::where([
                     ['producto_id', $datos['idproductos']],
                     ['unidades_id', $datos['idunidades']],
                     ['id_usuario', $idusuario],
-                    ['estado', 0]
-                ])->first();
-                if ($validacion) {
-                    $validacion->update([
-                        'estado' => 1
-                    ]);
+                    ['estado', 1]
+                ])->get();
+
+                if (count($validacionPublicacionExiste)) {
+                    $aErrores[] = '-Esta publicación ya existe';
                 } else {
 
-                    $nuevoPublicacion2 = new Publicaciones();
-                    $nuevoPublicacion2->precios_id = $datos['idpreciovendedor'];
-                    $nuevoPublicacion2->producto_id = $datos['idproductos'];
-                    $nuevoPublicacion2->unidades_id = $datos['idunidades'];
-                    $nuevoPublicacion2->id_usuario = $idusuario;
-                    $nuevoPublicacion2->estado = 1;
-                    $nuevoPublicacion2->descripcion = $datos['descripcion'];
-                    $nuevoPublicacion2->iva = $datos['iva'];
-                    $nuevoPublicacion2->created_at = \Carbon\Carbon::now();
-                    $nuevoPublicacion2->updated_at = \Carbon\Carbon::now();
-                    $nuevoPublicacion2->save();
-
-                    $validacionPrecio = Precios::where([
+                    $validacionUnidadProducto2 = Precios::where([
                         ['producto_id', $datos['idproductos']],
                         ['unidades_id', $datos['idunidades']],
-                        // ['id_asociacion', $idasociacion],
                         ['id_usuario', $idusuario],
-                    ])->first();
-                    if ($validacionPrecio) {
-                        $validacionPrecio->update([
-                            'precio' => $datos['precio'],
-                            'estado' => 1
-                        ]);
-                    }
-
-                    if (empty($_FILES['imagen']['name'][0])) {
-                        $productoImagen = Productos::find($datos['idproductos']);
-                        $imagenPredeterminada = $productoImagen->imagen;
-                        if ($imagenPredeterminada) {
-                            $nuevaImagenP = new Imagenes();
-                            $nuevaImagenP->ruta = $imagenPredeterminada;
-                            $nuevaImagenP->publicaciones_id = $nuevoPublicacion2->id;
-                            $nuevaImagenP->save();
-                        } else {
-                            $aErrores[] = '- Escoja al menos una imagen para la publicación o seleccione un producto con imagen predeterminada.';
-                        }
+                    ])->get();
+                    if (count($validacionUnidadProducto2)) {
+                        $aErrores[] = '-El precio de este producto ya está asignado a esta unidad';
                     } else {
-                        if (isset($datos['imagen']) && is_array($datos['imagen'])) {
-                            foreach ($datos['imagen'] as $imagen) {
-                                // Generar un nombre único para la imagen comprimida
-                                $nombreImagenComprimida = uniqid() . '.webp';
 
-                                // Construir la ruta para la imagen comprimida
-                                $rutaImagenComprimida = 'public/publicaciones/' . $nombreImagenComprimida;
+                        $validacion = Publicaciones::where([
+                            ['producto_id', $datos['idproductos']],
+                            ['unidades_id', $datos['idunidades']],
+                            ['id_usuario', $idusuario],
+                            ['estado', 0]
+                        ])->first();
+                        if ($validacion) {
+                            $validacion->update([
+                                'precio' => $datos['precio'],
+                                'estado' => 1
+                            ]);
+                        } else {
 
-                                // Cargar la imagen original y guardarla comprimida
-                                Image::load($imagen->getRealPath())
-                                    ->width(400)
-                                    ->height(400)
-                                    ->optimize()
-                                    ->save(storage_path("app/{$rutaImagenComprimida}"), 80, 'webp');
+                            $nuevoPublicacion2 = new Publicaciones();
+                            $nuevoPublicacion2->precios_id = $datos['idpreciovendedor'];
+                            $nuevoPublicacion2->producto_id = $datos['idproductos'];
+                            $nuevoPublicacion2->unidades_id = $datos['idunidades'];
+                            $nuevoPublicacion2->id_usuario = $idusuario;
+                            $nuevoPublicacion2->estado = 1;
+                            $nuevoPublicacion2->descripcion = $datos['descripcion'];
+                            $nuevoPublicacion2->iva = $datos['iva'];
+                            $nuevoPublicacion2->created_at = \Carbon\Carbon::now();
+                            $nuevoPublicacion2->updated_at = \Carbon\Carbon::now();
+                            $nuevoPublicacion2->save();
 
-                                // Obtener la URL de la imagen comprimida
-                                $urlImagen = Storage::url($rutaImagenComprimida);
+                            $validacionPrecio = Precios::where([
+                                ['producto_id', $datos['idproductos']],
+                                ['unidades_id', $datos['idunidades']],
+                                // ['id_asociacion', $idasociacion],
+                                ['id_usuario', $idusuario],
+                            ])->first();
+                            if ($validacionPrecio) {
+                                $validacionPrecio->update([
+                                    'precio' => $datos['precio'],
+                                    'estado' => 1
+                                ]);
+                            }
 
-                                $nuevaImagen = new Imagenes();
-                                $nuevaImagen->ruta = $urlImagen;
-                                $nuevaImagen->publicaciones_id = $nuevoPublicacion2->id;
-                                $nuevaImagen->save();
+                            if (empty($_FILES['imagen']['name'][0])) {
+                                $productoImagen = Productos::find($datos['idproductos']);
+                                $imagenPredeterminada = $productoImagen->imagen;
+                                if ($imagenPredeterminada) {
+                                    $nuevaImagenP = new Imagenes();
+                                    $nuevaImagenP->ruta = $imagenPredeterminada;
+                                    $nuevaImagenP->publicaciones_id = $nuevoPublicacion2->id;
+                                    $nuevaImagenP->save();
+                                }
+                                // else {
+                                //     $aErrores[] = '- Escoja al menos una imagen para la publicación o seleccione un producto con imagen predeterminada.';
+                                // }
+                            } else {
+                                if (isset($datos['imagen']) && is_array($datos['imagen'])) {
+                                    foreach ($datos['imagen'] as $imagen) {
+                                        // Generar un nombre único para la imagen comprimida
+                                        $nombreImagenComprimida = uniqid() . '.webp';
+
+                                        // Construir la ruta para la imagen comprimida
+                                        $rutaImagenComprimida = 'public/publicaciones/' . $nombreImagenComprimida;
+
+                                        // Cargar la imagen original y guardarla comprimida
+                                        Image::load($imagen->getRealPath())
+                                            ->width(400)
+                                            ->height(400)
+                                            ->optimize()
+                                            ->save(storage_path("app/{$rutaImagenComprimida}"), 80, 'webp');
+
+                                        // Obtener la URL de la imagen comprimida
+                                        $urlImagen = Storage::url($rutaImagenComprimida);
+
+                                        $nuevaImagen = new Imagenes();
+                                        $nuevaImagen->ruta = $urlImagen;
+                                        $nuevaImagen->publicaciones_id = $nuevoPublicacion2->id;
+                                        $nuevaImagen->save();
+                                    }
+                                }
                             }
                         }
                     }
+                    // if (count($aErrores) > 0) {
+                    //     $respuesta = array(
+                    //         'mensaje'      => $aErrores,
+                    //         'estado'      => 0,
+                    //     );
+                    //     return response()->json($respuesta);
+                    // } else {
+                    //     DB::commit();
+                    //     $respuesta = array(
+                    //         'mensaje'      => "",
+                    //         'estado'      => 1,
+                    //     );
+                    //     return response()->json($respuesta);
+                    // }
                 }
 
                 if (count($aErrores) > 0) {
@@ -347,9 +386,9 @@ class PublicacionesController extends Controller
         if ($datos['precio'] == "") {
             $aErrores[] = '- Digite el precio del producto según la unidad';
         }
-        if (empty($_FILES['imagen']['name'][0])) {
-            $aErrores[] = '- Escoja al menos una imagen para la publicación';
-        }
+        // if (empty($_FILES['imagen']['name'][0])) {
+        //     $aErrores[] = '- Escoja al menos una imagen para la publicación';
+        // }
         if (count($aErrores) > 0) {
             throw new \Exception(join('</br>', $aErrores));
         }
@@ -373,45 +412,67 @@ class PublicacionesController extends Controller
             $actualizarPublicacion->iva = $datos['iva'];
             $actualizarPublicacion->save();
 
+            //rutas de las imagenes
             $rutasImagenes = $actualizarPublicacion->imagenes()->pluck('ruta')->toArray();
+
+            //ruta imagen predeterminada
+            $productoImagen = Productos::find($datos['idproductos']);
+            $imagenPredeterminada = $productoImagen->imagen;
 
             // Eliminar físicamente las imágenes del servidor
             foreach ($rutasImagenes as $rutaImagen) {
-                // Completar la ruta completa según tu configuración
-                $rutaCompleta = public_path($rutaImagen);
 
-                // Verificar que el archivo exista antes de intentar eliminarlo
-                if (file_exists($rutaCompleta)) {
-                    unlink($rutaCompleta);
+                if ($rutaImagen !== $imagenPredeterminada) {
+                    $rutaCompleta = public_path($rutaImagen);
+
+                    // Verificar que el archivo exista antes de intentar eliminarlo
+                    if (file_exists($rutaCompleta)) {
+                        unlink($rutaCompleta);
+                    }
                 }
             }
 
             $actualizarPublicacion->imagenes()->delete();
 
-            // Agregar las nuevas imágenes
-            if (isset($datos['imagen']) && is_array($datos['imagen'])) {
-                foreach ($datos['imagen'] as $imagen) {
-                    // Generar un nombre único para la imagen comprimida
-                    $nombreImagenComprimida = uniqid() . '.webp';
 
-                    // Construir la ruta para la imagen comprimida
-                    $rutaImagenComprimida = 'public/publicaciones/' . $nombreImagenComprimida;
+            if (empty($_FILES['imagen']['name'][0])) {
+                $productoImagen2 = Productos::find($datos['idproductos']);
+                $imagenPredeterminada2 = $productoImagen2->imagen;
+                if ($imagenPredeterminada2) {
+                    $nuevaImagenP2 = new Imagenes();
+                    $nuevaImagenP2->ruta = $imagenPredeterminada2;
+                    $nuevaImagenP2->publicaciones_id = $actualizarPublicacion->id;
+                    $nuevaImagenP2->save();
+                }
+                // else {
+                //     $aErrores[] = '- Escoja al menos una imagen para la publicación o seleccione un producto con imagen predeterminada.';
+                // }
+            } else {
+                // Agregar las nuevas imágenes
+                if (isset($datos['imagen']) && is_array($datos['imagen'])) {
+                    foreach ($datos['imagen'] as $imagen) {
+                        // Generar un nombre único para la imagen comprimida
+                        $nombreImagenComprimida = uniqid() . '.webp';
 
-                    // Cargar la imagen original y guardarla comprimida
-                    Image::load($imagen->getRealPath())
-                        ->width(400)
-                        ->height(400)
-                        ->optimize()
-                        ->save(storage_path("app/{$rutaImagenComprimida}"), 80, 'webp');
+                        // Construir la ruta para la imagen comprimida
+                        $rutaImagenComprimida = 'public/publicaciones/' . $nombreImagenComprimida;
 
-                    // Obtener la URL de la imagen comprimida
-                    $urlImagen = Storage::url($rutaImagenComprimida);
+                        // Cargar la imagen original y guardarla comprimida
+                        Image::load($imagen->getRealPath())
+                            ->width(400)
+                            ->height(400)
+                            ->optimize()
+                            ->save(storage_path("app/{$rutaImagenComprimida}"), 80, 'webp');
 
-                    // Crear y guardar la nueva imagen asociada a la publicación
-                    $nuevaImagen = new Imagenes();
-                    $nuevaImagen->ruta = $urlImagen;
-                    $nuevaImagen->publicaciones_id = $actualizarPublicacion->id;
-                    $nuevaImagen->save();
+                        // Obtener la URL de la imagen comprimida
+                        $urlImagen = Storage::url($rutaImagenComprimida);
+
+                        // Crear y guardar la nueva imagen asociada a la publicación
+                        $nuevaImagen = new Imagenes();
+                        $nuevaImagen->ruta = $urlImagen;
+                        $nuevaImagen->publicaciones_id = $actualizarPublicacion->id;
+                        $nuevaImagen->save();
+                    }
                 }
             }
 
@@ -491,6 +552,32 @@ class PublicacionesController extends Controller
         }
         try {
             $eliminarPublicacion = Publicaciones::findOrFail($datos['id']);
+
+            // Obtener las rutas de las imágenes asociadas a la publicación
+            $rutasImagenes = $eliminarPublicacion->imagenes()->pluck('ruta')->toArray();
+
+            // Obtener la ruta de la imagen predeterminada del producto
+            $productoImagen = Productos::find($eliminarPublicacion->producto_id);
+            $imagenPredeterminada = $productoImagen->imagen;
+
+            // Eliminar las imágenes asociadas a la publicación actual, excluyendo la imagen predeterminada
+            foreach ($rutasImagenes as $rutaImagen) {
+                // Verificar si la ruta pertenece a la imagen predeterminada
+                if ($rutaImagen !== $imagenPredeterminada) {
+                    // Completar la ruta completa según tu configuración
+                    $rutaCompleta = public_path($rutaImagen);
+
+                    // Verificar que el archivo exista antes de intentar eliminarlo
+                    if (file_exists($rutaCompleta)) {
+                        unlink($rutaCompleta);
+                    }
+                }
+            }
+
+            $eliminarPublicacion->imagenes()->delete();
+            // Eliminar las imágenes asociadas a la publicación actual, excluyendo la imagen predeterminada
+            // $eliminarPublicacion->imagenes()->where('ruta', '!=', $imagenPredeterminada)->delete();
+
             $eliminarPublicacion->update(['estado' => 0]);
             DB::commit();
             $respuesta = array(
